Server async ops,Num_sects<2^12,Num_msgs_sect<2^16
Fewer messages passed between workers and servers. Type of message and other info in tag.
The tag is assumed to be 32 bit (asserts in place to ensure this).
4 bits are assigned to the message type.
12 bits to the section number (which pardo section in the program).
16 bits to the message number in each section.
If more sections or more messages per section are needed, we can repack the data differently or deal with this problem differently.