Added sqlite sources, wrapper & tests for wrapper.
Amalgamated sources for sqlite3 have been added and the build system has
been modified to accomodate for it. Extra flags have been inserted so
that sqlite3 does not require pthreads or dl libraries.

A wrapper class - ProfileTimerStore has been added to store information
collected by the profile interpreter. Data is stored in 6 separate
tables. Tables are made for 0 block operands, 1 block operand, 2 block
operands ... upto 5 block operands. Each table contains the opcode (as a
string), indices for each block, segment sizes for each block, the total
time spent on that super instruction and the number of times(count) that super
instruction was invoked.
The shape of the operation is captured by the indices of the blocks.
The shape of the blocks is captured by the segment sizes (or extents).
Each recorded time is for a given opcode with a given shape of operation
and shape of block operands.
This data store will be used for the modeling interpreter.

Tests have been added that check creation of tables, insertions &
deletions.