sqlite3 shell to cmake build. Attached timer store.
The sqlite3 "shell" is now built as part of the cmake build system.
This needs to be added to the autoconf/automake build system as well.
The "shell" is useful in examining the contents of the sqlite3 database
file generated by the profile interpreter run.

The ProfileTimerStore now saves upto 6 block operands.
The profile timer now also measures user defined super instructions.
Added tests to test the 6 block operands table.

The user defined super instruction, unlike the block operations, can
take as input operand - a static block without indices. The block
selector generated for this needs to be dealt with separately. Same goes
for scalars passed as parameters.

TODO :
1. The profile timer ends up saving any block wait time encountered.
This is not a problem for the single node version, but will be, when a
multinode version is added.
2. For now the name of the file generated is hard coded. This will change
to denote the job name & the mpi rank. For a multinode version, all
ranks after having generated their individual databases will need to be
merged somehow.