Updated the permissions (removed execute) to a few super instructions with +x
Moved MCPT fragment related super instructions to the sub directory mcpt-fragment

integral related routines not directly apart of the fragment code was moved to
direct_integral as well as fock matrix routines that control the integral calls.

some tidying of files, including depretiation of a number of sial and super
instructions.  depretiated files have an XXX ending and are removed from the
make system.