removed check that received block is the correct size.
This was a problem because this first implementation does not
distinguish between MPI_Request objects that are waiting for
a send to complete (where the only check of the status is for cancellation)
and an MPI_Request that is waiting for the ISend (which implements put)
to complete.  In the latter, checking for number of received bytes makes
no sense and causes a crash.

We may want to revisit this decision later.

Also, simplified and cleaned up.