Add Counter, MaxCounter, and SimpleTimer classes.  Tracer is now useful.
These classes allow per core timing and counts to be collected.  i
This is in contrast with the existing timer infrastructure that only
reports averages across cores.  These classes have not been designed for
scalability, but rather to give useful information about the inner workings
of the sip on reasonably small test cases.  They also will not compile without
MPI. Fixing this will be done after determining whether or not these classes
are worth keeping.

This commit introduces  asynchronous puts, using the regular block map (not the cached one).