this patch adds two new tests:   one additional ccsdpt (second_ccsdpt_test)   a new CIS(D) (cis_test)
with changes to the standard printing for easier use as a quantum chemist the
standard sial tests expected output had to be updated.  It is my hope this will
not have to be done again.

also added is the testing ability for static set to persistent.  this is
modestly important as quantities one would like to store are often the
quantities one would liek to test.