Preliminary implementation of enhancement of put to allow scalar rhs.
Constructs like

   tmp[i,j] = 0.0
   put a[i,j] = tmp[i,j]

can be replaced with
   put a[i,j] = 0.0

Also,
      put a[i,j] += 2.0
increments each element by 2,

and
    put a[i,j] *= -1.0
multiplies each element by -1.

In contrast to sending an entire block (typically filled with zeros), only
the value is sent and the server performs the initialization.

More tests and some cleanup are still needed.

A bug was fixed in ../src/sialx/qm/rcis_rhf.sialx.
A barrier was added inside the do loops at line 1627.
This is now required since it is no longer guaranteed that
the iterations of the pardo loop with the same values of a and i
will be executed on the same worker.