Commit for correct eom-ccsd
Test case works.

some notes:

-citations in the sialx file contain most of the useful information.

-DGEEV is a bit finiky with the BiGS algorithm because of small norms.  The
current workaround is the removal of crash on imaginary and a rescale to handle
the norm.  This does not seem to change anything numerically but care as usual
for excited states should be exercised.

-This commit is prior to the inclusion of the put_initialize branch which will
enable a significant change in the normalization code.

To run the eom code, simply add

SIAL_PROGRAM = lr_eom_ccsd_rhf.siox

after a CCSD job.  It expects the tran_rhf_no4v.siox to have been run.