preliminary implementation of worker side prefetching
The goal is to get an idea of the benefits before
spending more time on the implementation

Current Limitations:

Only prefetches when inner loop is a do loop.

Cannot handle if statements and where clauses in the inner loop.
Programs with this
may fail tests because worker will attempt to prefetch a block that will
not ever be requested, and thus may not have been initialized, causing an abort.
Two of the qm tests have been disabled due to this limitation.