server prioritizes short messages.  Does not handle same source gets correctly
This implementation is preliminary and does not handle the complete
sial semantics.  It will allow us to evaluate the performance impact
of handling the data messages for put and put accumulate instructions
separately from the first message.  The idea is to ensure that an
Irec has been posted for all large messages before the message is
sent, thus avoiding mpi buffering issues.

This commit prioritizes handling of short messages over
other put data and put+ data messages.  This uses more
memeory for temporary blocks at the server but should give
better response time for small messages.

A complete solution needs to ensure that all messages from a single worker
are handled in order.  This can be done by restricting sial semantics or
by changing the implementation at the server to keep track of this.