This patch addresses the error in EOM-CCSD for multiple segments.
The theory is that for multiple segments, partial sums for the Raa and HBaa
terms were incorrectly adapted to full sums in the subspace-davidson and L-R
orthogonalization.  Usual attempts to address this failed, so due to time
constraints direct calculation of Raa and HBaa is dropped.  Instead Raa and HBaa
is constructed from antisymmetrized Rab and HBab (and for the L vector too).
Effects should be limited to the definite inability to do triplets.  But not
much else?

Jason.