new index format for sparse persistentat arrays + restart
Two file formats for the index for persistent arrays are supported.
Dense includes an entry for each potential block in the array.
Sparse contains a block_number, offset pair for each actual block.

The choice between the two formats is currently hard code in the
is_dense variable in line 267 of disk_backed_block_map.cpp.  We leave for
later the task of making a good choice, probably using the sial sparse
keyword.

Each job constructs a job_id from the time.  Then the data for a persistent
distributed/served array is stored in a file job_id.label.program_number.parr with
its index data in job_id.array_label.program_number.parr_index.  During the
execution, temporary files named job_id.array_name.program_number.arr
are used for disk backing.  If the array is persistent, the file is renamed
at the end of that sial program, otherwise, it is deleted.

Persistent scalars and contiguous arrays at workers are written to a file called
job_id.program_number.worker_checkpoint by rank 0 at the end of the sial program.
If the job is restarted, the file is read to populate the in-memory data strutures
in WorkerPersistentArrayManager.

At the end of every sial program, the program number of the next program to run is
written to a file nameed job_id.log.

A job can be restarted by simply adding an additional parameter to the command line with
option b.

For example,

mpirun -np 64 ./aces4 -d my_job.dat  -s src/sialx/qm -b _2015Nov07_201652

The -b option will cause the log file to be read to obtain the number of the sial program to start.
During initialization, the worker_checkpoint file is loaded. The arrays are moved to the
workers normal data_manager after a restore command.  Blocks of distributed arrays are read when
the array is restored.

Restart does not yet work on real quantum chemistry code due to the common block defined in imemory.h in the
integral package.  Someone needs to write little routines to put and get those values from a file.
(This can be easily done using the same IO routines that are used for creating .dat files.

The GlobalState class which has only static members has been replaced with a class called
JobControl.  There is a static global pointer to a single JobContrl instance.  This was easier
to manage in the test suites.