Changed most instances of check, check_and_warn, etc. to macros
The problem using functions check(..) means that all parameters
are evaluated before the function is called.  Since many of them
require constructing and manipulating strings, this is unnecessary
work.  The macros check the condition, first, then construct
the error message only if it is necessary.

Also the definition of the macro can be easily changed to get rid
of the checks altogether.