TestControllerParallel has static sleep_between_jobs variable
Indicates the number of seconds to sleep between test cases.
Default is zero

This variable can be used to address failures in test suites when tests receive the
same jobid.  This occurs because the timestamp used to obtain jobids has low resolution.
In this commit, the value is set to 5 in test_qm.