added utility check_system.  Checks that the MPI tags will not exceed MPI_TAG_UB
check_system gets the system value for MPI_TAG_UB and the maximum value
for a tag that will be generated by BarrierSupport.  The latter is
compared with MPI_TAG_UB for the system it is run on, and the
minimum value in an MPI standard conforming implementation.

Eventually, other checks should be migrated to this program, which
only needs to be run once after installation of Aces4.